+++ editor/extensions/ext-star.js
@@ -42,7 +42,7 @@ export default {
 
     function setAttr (attr, val) {
       svgCanvas.changeSelectedAttribute(attr, val);
-      S.call('changed', selElems);
+      svgCanvas.call('changed', selElems);
     }
 
     /*
@@ -118,12 +118,12 @@ export default {
         if (svgCanvas.getMode() === 'star') {
           started = true;
 
-          newFO = S.addSVGElementFromJson({
+          newFO = svgCanvas.addSVGElementFromJson({
             element: 'polygon',
             attr: {
               cx: opts.start_x,
               cy: opts.start_y,
-              id: S.getNextId(),
+              id: svgCanvas.getNextId(),
               shape: 'star',
               point: document.getElementById('starNumPoints').value,
               r: 0,